snn_async_delays/requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Scientific computing
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0

# Config / utilities
pyyaml>=6.0
tqdm>=4.65.0
